implementation of ad and user controllers methods

+++ code/donors_club/src/main/java/pt/donors/club/donors_club/repositories/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import pt.donors.club.donors_club.models.Category;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
-    
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
+
 }
