+++ code/donors_club/src/main/java/pt/donors/club/donors_club/repositories/CityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import pt.donors.club.donors_club.models.City;
 
-public interface CityRepository extends CrudRepository<City, Long> {
-    
+public interface CityRepository extends CrudRepository<City, Integer> {
+
 }
