+++ code/donors_club/src/main/java/pt/donors/club/donors_club/repositories/DistrictRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import pt.donors.club.donors_club.models.District;
 
-public interface DistrictRepository extends CrudRepository<District, Long> {
-    
+public interface DistrictRepository extends CrudRepository<District, Integer> {
+
 }
