+++ code/donors_club/src/main/java/pt/donors/club/donors_club/repositories/MessageRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import pt.donors.club.donors_club.models.Message;
 
-public interface MessageRepository extends CrudRepository<Message, Long> {
+public interface MessageRepository extends CrudRepository<Message, Integer> {
 }
