+++ code/donors_club/src/main/java/pt/donors/club/donors_club/repositories/WishListRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import pt.donors.club.donors_club.models.WishList;
 
-public interface WishListRepository extends CrudRepository<WishList, Long> {
-    
+public interface WishListRepository extends CrudRepository<WishList, Integer> {
+
 }
